Diskovery adapter: fixed several possible type conversion problems
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@15699 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Diskovery/Diskovery.cpp
@@ -480,7 +480,7 @@ int DiskoveryStateDev::Initialize()
    hub_->GetLabel(hubLabel);
    SetParentID(hubLabel); // for backward compatibility (delete?)
 
-   for (unsigned int i = 0; i < numPos_; i++) {
+   for (uint16_t i = 0; i < numPos_; i++) {
       ostringstream os;
       os << "Preset-" << (i + firstPos_);
       // set default label
